+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added ProgressColumn `MofNCompleteColumn` to display raw `completed/total` column (similar to DownloadColumn,
+  but displays values as ints, does not convert to floats or add bit/bytes units). 
+  https://github.com/Textualize/rich/discussions/1899
+
+
 ## [11.2.0] - 2022-02-08
 
 ### Added